LayoutTests/http/tests/security/xssAuditor/javascript-link-HTML-entities-expected.txt

CONSOLE MESSAGE: line 14: Refused to execute a JavaScript script. Source code of script found within request.
